% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{analyze}
\alias{analyze}
\title{LONGO analysis}
\usage{
analyze(DATA.DF, MULTI_PROBES, WINDOW_SIZE, STEP_SIZE, WINDOW_STYLE, FILTERED,
  NORMALIZED, CONTROL_COLUMN_INDEX)
}
\arguments{
\item{DATA.DF}{dataframe to be analyzed}

\item{MULTI_PROBES}{Option to select what to do with multiple probes to a single gene, default=highest}

\item{WINDOW_SIZE}{Option to alter the size of the windows to be created from the dataframe, default=200}

\item{STEP_SIZE}{Option to alter the size of the steps used to create the windows, default=40}

\item{WINDOW_STYLE}{Option to alter the method used to create windows, default=mean}

\item{FILTERED}{Option to filter data, default=TRUE}

\item{NORMALIZED}{Option to normalize data, default=TRUE}

\item{CONTROL_COLUMN_INDEX}{column of the data used as a control for the statistical analyses}
}
\value{
returns the simplified data
}
\description{
This is a internal function. Used by the package to take a formatted dataframe and analyzes it based on the user inputted values.
}

