% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbiomaRt.r
\name{callbiomaRt}
\alias{callbiomaRt}
\title{callbiomaRt}
\usage{
callbiomaRt(data.df, library_type, ensembl)
}
\arguments{
\item{data.df}{dataframe that contains the probe ids that will be checked}

\item{library_type}{the gene identifier of the data based off the biomaRt database}

\item{ensembl}{BiomaRt mart based on the given species}
}
\value{
returns the data.df with two new columns, length and symbol
}
\description{
calls the biomaRt databse based on inputed data
}

