% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{dict}
\alias{dict}
\title{dict}
\usage{
dict(data.df, source_data.bm)
}
\arguments{
\item{data.df}{inputed data}

\item{source_data.bm}{biomaRt database}
}
\value{
returns a dictionary
}
\description{
creates and returns a dicitonary
}

