% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example dataset to show LONGO}
\format{A data frame with 45,101 rows and 19 variables:
\describe{
    \item{ID_REF}{identifier used for the biomaRt database}
    \item{GSM480852}{2}
    \item{GSM480853}{3}
    \item{GSM480854}{4}
    \item{GSM480855}{5}
    \item{GSM480856}{6}
    \item{GSM480857}{7}
    \item{GSM480858}{8}
    \item{GSM480859}{9}
    \item{GSM480860}{10}
    \item{GSM480861}{11}
    \item{GSM480862}{12}
    \item{GSM480863}{13}
    \item{GSM480864}{14}
    \item{GSM480865}{15}
    \item{GSM480866}{16}
    \item{GSM480867}{17}
    \item{GSM480868}{18}
    \item{GSM480869}{19}
}}
\usage{
exampleData
}
\value{
Returns nothing
}
\description{
A dataset to show expression values:
}
\keyword{datasets}

