% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleHumanData}
\alias{exampleHumanData}
\title{Example Human dataset}
\format{A data frame with 49,395 rows and 7 columns:
\itemize{
  \item{ID_REF}{1}
  \item{GSM1701831}{2}
  \item{GSM1701832}{3}
  \item{GSM1701833}{4}
  \item{GSM1701834}{5}
  \item{GSM1701835}{6}
  \item{GSM1701836}{7}
}}
\source{
\url{https://ftp.ncbi.nlm.nih.gov/geo/series/GSE69nnn/GSE69480/matrix/GSE69480_series_matrix.txt.gz}
}
\description{
Hu W, et al., Cell Stem Cell 2015
Raw data  was downloaded from: https://ftp.ncbi.nlm.nih.gov/geo/series/
GSE69nnn/GSE69480/matrix/GSE69480_series_matrix.txt.gz
The replicates for each sample were averaged to get a single sample for
 easy testing.
}
