% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{GOanalysis}
\alias{GOanalysis}
\title{LONGO GO analysis}
\usage{
GOanalysis(analyzeData, biomaRt.go.df, speciesEnsembl, repNames, GODomain,
  GOScoring, GoGraphing, sigCount = 15, sigCutOff = 0.01,
  textScale = 0.15)
}
\arguments{
\item{analyzeData}{dataframe to be analyzed}

\item{biomaRt.go.df}{dataframe of gene id to GO term}

\item{speciesEnsembl}{Ensembl used to call the bioMart if biomaRt.df is null}

\item{repNames}{Order/naming of the replications}

\item{GODomain}{Option for what GO domain to use for the analysis}

\item{GOScoring}{Option for what method should be used for the GO scoring}

\item{GoGraphing}{Option for what method should be used for the GO graphig}

\item{sigCount}{Option for the number of significant terms are needed in
the nodes to count}

\item{sigCutOff}{Option for the p value to mark significance}

\item{textScale}{Option to alter text size in GO graph}
}
\value{
returns the GO data
}
\description{
This is an internal function used to do the GO analysis.
}
