% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LONGO.r
\name{LONGO}
\alias{LONGO}
\title{Use LONGO through shiny interface}
\usage{
LONGO()
}
\description{
This function allows the user to input data files and alter the input variables to make sure the formatting is correct.
They can then run thte LONGO package which will output the results and plots in the browser and allow the user to download results as needed.
}

