% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LONGOquotient.R
\name{LONGOquotient}
\alias{LONGOquotient}
\title{Calculates LONGO quotient}
\usage{
LONGOquotient(data.df, JSdata.df, controlColumn)
}
\arguments{
\item{data.df}{The simplified data from analysis}

\item{JSdata.df}{The JS distance data from analysis}

\item{controlColumn}{The column in the data which should be used as
control}
}
\value{
Returns nothing
}
\description{
This is an internal functuon. This function calculcates the LONGO quotient.
}

