% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfile.r
\name{openfile}
\alias{openfile}
\title{File opener}
\usage{
openfile(FILE, SEP, HEADER, COMMENT_CHAR)
}
\arguments{
\item{FILE}{location to the file to be opened}

\item{SEP}{variable that is used to separate columns}

\item{HEADER}{true/false used to open a file with or without a header}

\item{COMMENT_CHAR}{character used as commented lines to be ignored in the file}
}
\value{
Returns nothing
}
\description{
This is an internal functuon. This function opens a file
and reads the data into a dataframe.
}

