% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLONGO.r
\name{plotLONGO}
\alias{plotLONGO}
\title{main function}
\usage{
plotLONGO(simp.df)
}
\arguments{
\item{simp.df}{dataframe to be plotted}
}
\description{
analyzes data by creating windows and moving by a step size
}

