% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{analyze}
\alias{analyze}
\title{analyze data
Takes in a dataframe analyzes it. internal function
 @param data.df dataframe to be analyzed
 @param multi_probes Option to select what to do with multiple probes to a single gene, default=highest
 @param window_size Option to alter the size of the windows to be created from the dataframe, default=200
 @param step_size Option to alter the size of the steps used to create the windows, default=40
 @param window_style Option to alter the method used to create windows, default=mean
 @param filtered Option to filter data, default=TRUE
 @param normalized Option to normalize data, default=TRUE
 @param control_column_index column of the data used as a control for the statistical analyses
 @importFrom edgeR cpm
 @importFrom preprocessCore normalize.quantiles
 @importFrom data.table as.data.table
 @importFrom data.table as.data.frame
 @importFrom data.table .SD
 @return returns the simplified data}
\usage{
analyze(data.df, multi_probes, window_size, step_size, window_style, filtered,
  normalized, control_column_index)
}
\description{
analyze data
Takes in a dataframe analyzes it. internal function
 @param data.df dataframe to be analyzed
 @param multi_probes Option to select what to do with multiple probes to a single gene, default=highest
 @param window_size Option to alter the size of the windows to be created from the dataframe, default=200
 @param step_size Option to alter the size of the steps used to create the windows, default=40
 @param window_style Option to alter the method used to create windows, default=mean
 @param filtered Option to filter data, default=TRUE
 @param normalized Option to normalize data, default=TRUE
 @param control_column_index column of the data used as a control for the statistical analyses
 @importFrom edgeR cpm
 @importFrom preprocessCore normalize.quantiles
 @importFrom data.table as.data.table
 @importFrom data.table as.data.frame
 @importFrom data.table .SD
 @return returns the simplified data
}

