% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{analyze}
\alias{analyze}
\title{LONGO analysis}
\usage{
analyze(analyzeData, multiProbes, windowSize, stepSize, windowStyle, filterData,
  normalizeData, controlColumn)
}
\arguments{
\item{analyzeData}{dataframe to be analyzed}

\item{multiProbes}{Option to select what to do with multiple probes to a
single gene, default=mean}

\item{windowSize}{Option to alter the size of the windows to be created
from the dataframe, default=200}

\item{stepSize}{Option to alter the size of the steps used to create the
windows, default=40}

\item{windowStyle}{Option to alter the method used to create windows,
default=mean}

\item{filterData}{Option to filter data using edgeR::cpm, default=TRUE}

\item{normalizeData}{Option to normalize data using
preprocessCore::normalize.quantiles, default=TRUE}

\item{controlColumn}{column of the data used as a control for the
statistical analyses}
}
\value{
returns the simplified data
}
\description{
This is a internal function. Used by the package to take a formatted
dataframe and analyzes it based on the user inputted values.
}
