% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbiomaRt.r
\name{callbiomaRt}
\alias{callbiomaRt}
\title{LONGO call biomaRt}
\usage{
callbiomaRt(data.df, libraryType, ensembl)
}
\arguments{
\item{data.df}{dataframe that contains the probe ids that will be checked}

\item{libraryType}{the gene identifier of the data based off the
biomaRt database}

\item{ensembl}{BiomaRt mart based on the given species}
}
\value{
returns the data.df with two new columns, length and symbol
}
\description{
This is a internal function. This function calls the biomaRt database to
retrieve the gene symbols and lengths.
}
