% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{dict}
\alias{dict}
\title{LONGO dictionary function}
\usage{
dict(data.df, source_data.bm)
}
\arguments{
\item{data.df}{inputed data}

\item{source_data.bm}{biomaRt database}
}
\value{
returns a dictionary
}
\description{
This is a internal function. This function creates and returns a dictionary
based on the user inputted values. Used to quickly link expression data
with names and lengths.
}

