% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LONGOcmd.R
\name{LONGOcmd}
\alias{LONGOcmd}
\title{LONGO function}
\usage{
LONGOcmd(fileLocation, separator = ",", header = TRUE, commentChar = "!",
  species = "hsapiens_gene_ensembl", libraryType = "affy_primeview",
  multiProbes = "mean", windowSize = 200, stepSize = 40,
  windowStyle = "mean", filterData = TRUE, normalizeData = TRUE,
  controlColumn = 2)
}
\arguments{
\item{fileLocation}{Address to the file to be analyzed}

\item{separator}{Option to specify the separated value, default=","}

\item{header}{Option to specify if there is a header in the file,
default=TRUE}

\item{commentChar}{Option to specify the commented character in the loaded
file, default="!"}

\item{species}{species name of the data based off the biomaRt database}

\item{libraryType}{the gene identifier of the data based off the
biomaRt database}

\item{multiProbes}{Option to change the method for handling multiple
probes going to the same gene, default=mean}

\item{windowSize}{Option to alter the size of the windows to be created
from the dataframe, default=200}

\item{stepSize}{Option to alter the size of the steps used to create the
windows, default=40}

\item{windowStyle}{Option to alter the method used to create windows,
default=mean}

\item{filterData}{Option to filter data, default=TRUE}

\item{normalizeData}{Option to normalize data, default=TRUE}

\item{controlColumn}{column of the data used as a control for the
statistical analysis}
}
\value{
Returns nothing
}
\description{
This function allows the user to preset the variables and allows a faster
analysis of data. Please use the shiny interface first to get an
understanding of how this package analyzes data.
}
\examples{
LONGOcmd(exampleData, species = "mmusculus_gene_ensembl",
libraryType = "affy_mouse430a_2")
\donttest{
LONGOcmd("~/data.csv")
LONGOcmd("~/data.csv", windowSize=100, stepSize=20)
LONGOcmd("~/data.csv", multiProbes="highest")
}
}
