% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfile.R
\name{openfile}
\alias{openfile}
\title{File opener}
\usage{
openfile(fileLocation, separator, header, commentChar)
}
\arguments{
\item{fileLocation}{location to the file to be opened}

\item{separator}{variable that is used to separate columns}

\item{header}{true/false used to open a file with or without a header}

\item{commentChar}{character used as commented lines to be ignored
in the file}
}
\value{
Returns nothing
}
\description{
This is an internal functuon. This function opens a file
and reads the data into a dataframe.
}

