% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLONGO.r
\name{plotLONGO}
\alias{plotLONGO}
\title{Plot function}
\usage{
plotLONGO(simp.df, x_label, y_label)
}
\arguments{
\item{simp.df}{dataframe to be plotted}

\item{x_label}{x label for plot}

\item{y_label}{y label for plot}
}
\value{
Returns a plot
}
\description{
This is a internal function. This function allows the program to
plot the data.
}
