% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLONGO.r
\name{plotLONGO}
\alias{plotLONGO}
\title{Plot function}
\usage{
plotLONGO(simp.df)
}
\arguments{
\item{simp.df}{dataframe to be plotted}
}
\description{
This is a internal function. This function allows the program to plot the data.
}

